- Small improvements and refactors
